requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
apispec==6.6.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
dataclasses-json==0.6.6
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
flake8==7.1.1
Flask==3.0.3
Flask-DB==0.4.1
Flask-Migrate==4.0.7
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-experimental==0.0.64
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langfuse==2.33.1
langsmith==0.1.106
llama-index==0.10.39
llama-index-agent-openai==0.2.5
llama-index-callbacks-langfuse==0.1.4
llama-index-cli==0.1.12
llama-index-core==0.10.39.post1
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.21
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-pinecone==0.1.7
llama-parse==0.4.3
llamaindex-py-client==0.1.19
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.42.0
orjson==3.10.7
packaging==23.2
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
pinecone-client==3.2.2
platformdirs==4.2.2
pluggy==1.5.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-service-object==0.1.0
pyarrow==16.1.0
pycodestyle==2.12.1
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pypdf==4.2.0
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
streamlit==1.35.0
striprtf==0.0.26
tenacity==8.3.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
twilio==9.2.4
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
webargs==8.4.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
